pyproject.toml
[build-system]
requires = [ "setuptools" ]
build-backend = 'setuptools.build_meta'

[tool.setuptools.packages.find]
include = ["terratorch*"]

[project]
name = "terratorch"
version = "v1.2.4"
description = "TerraTorch - The geospatial foundation model fine-tuning toolkit"
license = { "text" = "Apache License, Version 2.0" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["fine-tuning", "geospatial foundation models", "artificial intelligence"]
authors = [
  { email = "terratorch@ibm.com"},
  { name = "Carlos Gomes"},
  { name = "Joao Lucas de Sousa Almeida"},
  { name = "Romeo Kienzler"},
  { name = "Benedikt Blumenstiel"},
  { name = "Francesc Martí Escofet"},
  { name = "Paolo Fraccaro"},
  { name = "Pedro Henrique Conrado"},
  { name = "Jaione Tirapu Azpiroz"},
  { name = "Daniela Szwarcman"},
  { name = "Bianca Zadrozny"},
  { name = "Þorsteinn Elí Gíslason"},
  { name = "Raunak Bhansali"},
  { name = "Takao Moriyama"},
  { name = "Þorsteinn Elí Gíslason"},
  { name = "Isabelle Wittmann"}
]
classifiers = [
#  "Development Status :: 2 - Pre-Alpha",
#  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
#  "Programming Language :: Python :: Implementation :: CPython",
]

dependencies = [
  "torch>2.0",
  "numpy>=2.2",
  "torchvision",
  "rioxarray",
  "albumentations",
  "albucore", 
  "rasterio",
  "torchmetrics",
  "geopandas",
  "lightly==1.5.22",
  "h5py",
  "lightning>=2.6.0",
  "segmentation-models-pytorch>=0.5.0",
  "jsonargparse>=4.40.0",
  "torchgeo", #>=0.7.0,<0.7.2",
  "einops",
  "timm>=1.0.15",
  "pycocotools",
  "huggingface_hub",
  "tifffile",
  "python-box",
  "tqdm",
  "wandb>=0.18.0",
  "tacoreader==0.5.6",
  "tensorboard",
  "diffusers",
  "scikit-learn>=1.3.2",
  "scikit-image",
  "pyarrow",
  "rich",
  "termcolor"
]


[project.entry-points."vllm.io_processor_plugins"]
terratorch_segmentation = "terratorch.vllm.plugins.segmentation:register_segmentation_plugin"
terratorch_tm_segmentation = "terratorch.vllm.plugins.segmentation:register_terramind_segmentation_plugin"

[project.optional-dependencies]
dev = [
  "black",
  "mkdocs-material",
  "mkdocstrings[python]",
  "mike", # for building docs with versions
  "mkdocs-git-revision-date-localized-plugin",
  "pre-commit"
]

test = [
  "pytest",
  "pytest-coverage",
  "coverage-badge",
  "peft>=0.15.0",
  "diffusers",
  "tokenizers",
  "dask",
  "numba",
  "hdf5plugin",
  "h5netcdf",
  "geobenchv2>=0.9",
  "terratorch-surya>=0.1.0",
#  "PrithviWxC",  # Currently fails of pypi version conflics
#  "granitewxc",  # Currently fails of pypi version conflics
]

surya = [
  "numba",
  "hdf5plugin",
  "h5netcdf",
  "terratorch-surya==0.1.0"
]

mmseg = [
  "mmcv",
  "ftfy",
  "regex",
  "openmim"
]

wxc = [
  "PrithviWxC",
  "granitewxc"
]

visualize = [
  "torchview",
  "graphviz"
]

peft = [
  "peft>=0.15.0"
]

geobenchv2 = [
  "geobenchv2==0.9",
  "mlflow",
]

vllm = [
  "geobenchv2>=0.9",
  "impactmesh>0.1.0",
  "vllm>=0.12,!=0.15.*",
]

vllm_test = [
  "pillow",
  "imagehash",
]

[project.urls]
Documentation = "https://terrastackai.github.io/terratorch/"
Issues = "https://github.com/terrastackai/terratorch/issues"
Source = "https://github.com/terrastackai/terratorch"

[project.scripts]
terratorch = "terratorch.__main__:main"

[[tool.mypy.overrides]]
module = "timm"
ignore_missing_imports = true

[tool.pytest.ini_options]
norecursedirs = [".venv", "build", "dist", "site-packages"]

[tool.black]
target-version = ["py310"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py310"
line-length = 120
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["terratorch"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
# Tasks rely on torch lightning saving args to hparams dict
"terratorch/tasks/*" = ["ARG002"]

[tool.coverage.run]
source_pkgs = ["terratorch", "tests"]
branch = true
parallel = true
omit = [
  "terratorch/__about__.py",
]

[tool.coverage.paths]
terratorch = ["terratorch"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.bumpver]
current_version = "0.99.5"
version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
